// Soe.Angular.Spa/src/app/shared/models/generated-interfaces/Enumerations.ts
export enum MatrixDataType { 
    String = 1,
    Integer = 2,
    Boolean = 3,
    Date = 4,
    Decimal = 5,
    Time = 6,
    DateAndTime = 7
}
// ...
export enum TermGroup_MatrixGroupAggOption { 
    Sum = 1,
    Min = 2,
    Max = 3,
    Count = 4,
    Average = 5,
    Median = 6,
    None = 100
}

export enum TermGroup_MatrixDateFormatOption { 
    DateShort = 1,
    DateLong = 2,
    DayOfMonth = 11,
    DayOfMonthPadded = 12,
    DayOfYear = 21,
    WeekOfYear = 31,
    DayOfWeekShort = 41,
    DayOfWeekFull = 42,
    Month = 51,
    MonthPadded = 52,
    NameOfMonthShort = 53,
    NameOfMonthFull = 54,
    Quarter = 61,
    YearShort = 71,
    YearFull = 72,
    YearMonth = 81,
    YearWeek = 91
}