// Soe.Angular.Spa/src/app/shared/models/generated-interfaces/MatrixResult.ts
// This file was generated by TypeWriter. See TypeWriter.Common.tst for reference.



    export interface IMatrixDefinitionColumn {
        columnNumber: number;
        key: string;
        matrixDataType: MatrixDataType;
        field: string;
        title: string;
        options: IMatrixDefinitionColumnOptions;
        matrixLayoutColumn: IMatrixLayoutColumn;
    }

    

    

    export interface IMatrixDefinitionColumnOptions {
        hidden: boolean;
        key: string;
        alignLeft: boolean;
        alignRight: boolean;
        clearZero: boolean;
        changed: boolean;
        decimals: number;
        minutesToDecimal: boolean;
        minutesToTimeSpan: boolean;
        formatTimeWithSeconds: boolean;
        formatTimeWithDays: boolean;
        dateFormatOption: TermGroup_MatrixDateFormatOption;
        labelPostValue: string;
        groupBy: boolean;
        aggregate: boolean;
        groupOption: TermGroup_MatrixGroupAggOption;
    }

    

    

    export interface IMatrixLayoutColumn {
        matrixDataType: MatrixDataType;
        field: string;
        title: string;
        sort: number;
        visible: boolean;
        options: IMatrixDefinitionColumnOptions;
    }

    

    




export enum MatrixTableType { 
    Unkown = 0,
    TimePayrollTransactionDTO = 501,
    TimePayrollScheduleTransactionDTO = 502
}



import { MatrixDataType, TermGroup_MatrixGroupAggOption, TermGroup_MatrixDateFormatOption } from "../generated-interfaces/Enumerations";